[FE] refactor: router 관련 함수 따로 분리
